// dafne/src/app/models/synchronizer.ts
export class Synchronizer {
    id: number;
    label: string;
    serviceUrlBackend: string;
    serviceUrl: string;
    serviceLogin: string;
    servicePassword: string;
    copyProduct: string;
    schedule: string;
    pageSize: number;
    request: string;
    targetCollection: any;
    remoteIncoming: string;
    sourceCollection: string;
    lastCreationDate: string;
    filterParam: string;
    geoFilter: string;
    status: string;
    skipOnError: string;
}